Issue 75

+++ src/main/java/com/github/wasiqb/coteafs/appium/config/PlaybackSetting.java
@@ -25,7 +25,7 @@ public class PlaybackSetting {
 	private int				defaultWait;
 	private int				delayAfterSwipe;
 	private int				delayAfterTap;
-	private int				delayBeforeSwipe;
+	private long			delayBeforeSwipe;
 	private int				delayBeforeTap;
 	private RecordSetting	record;
 	private boolean			screenshotOnError;
@@ -82,7 +82,7 @@ public class PlaybackSetting {
 	 * @since Jan 18, 2018 9:35:15 PM
 	 * @return the delayBeforeSwipe
 	 */
-	public int getDelayBeforeSwipe () {
+	public long getDelayBeforeSwipe () {
 		return this.delayBeforeSwipe;
 	}
 
@@ -176,7 +176,7 @@ public class PlaybackSetting {
 	 * @param delayBeforeSwipe
 	 *            the delayBeforeSwipe to set
 	 */
-	public void setDelayBeforeSwipe (final int delayBeforeSwipe) {
+	public void setDelayBeforeSwipe (final long delayBeforeSwipe) {
 		this.delayBeforeSwipe = delayBeforeSwipe;
 	}
